Update radio.c
With your permission, I've slightly modified the radio station list. Adjusted frequencies. Changed/added new names. I'm relatively new here, so I don't know all the intricacies of working with apps and code adjustments. But I'm a radio amateur and I'm interested in using Flipper in this form. I'd be very grateful if my contribution adds convenience and a new experience to someone else. In general, if I were you, I'd change the automatic search to manual, with a 100 kHz step. Because the sensitivity of the TEA5767 radio module is very poor. Best regards, Aleks Y.